double to float

+++ include/wolf/math.h
@@ -18,8 +18,8 @@
 
 typedef struct	s_v2
 {
-	double		x;
-	double		y;
+	float		x;
+	float		y;
 }				t_v2;
 
 typedef struct	s_hit
@@ -34,23 +34,23 @@ typedef struct	s_line
 	t_v2		b;
 }				t_line;
 
-inline int		math_fl(double x);
-inline int		math_cl(double x);
-inline int		math_isfl(double x);
-inline double	math_dec(double x);
-inline t_v2		v2_turn(t_v2 a, double t);
+inline int		math_fl(float x);
+inline int		math_cl(float x);
+inline int		math_isfl(float x);
+inline float	math_dec(float x);
+inline t_v2		v2_turn(t_v2 a, float t);
 inline t_v2		v2_rag(t_v2 a);
 inline t_v2		v2_sub(t_v2 a, t_v2 b);
 inline t_v2		v2_add(t_v2 a, t_v2 b);
-inline t_v2		v2_mul(t_v2 a, double n);
-inline double	v2_mag(t_v2 a);
+inline t_v2		v2_mul(t_v2 a, float n);
+inline float	v2_mag(t_v2 a);
 inline t_v2		v2_unit(t_v2 a);
-inline double	v2_slope(t_v2 a);
+inline float	v2_slope(t_v2 a);
 inline t_v2		v2_sh(t_v2 a, t_v2 b);
 inline t_v2		v2_sv(t_v2 a, t_v2 b);
 inline int		v2_tile(t_v2 a, const char **tiles);
-inline double	v2_pcast(double size, int yres, int y);
-inline t_line	line_rotate(t_line l, double t);
-inline t_v2		line_lerp(t_line l, double n);
+inline float	v2_pcast(float size, int yres, int y);
+inline t_line	line_rotate(t_line l, float t);
+inline t_v2		line_lerp(t_line l, float n);
 
 #endif
